[feat] progressView자연스럽게 하려다가 그냥 원상복귀..
