+++ PIM/PIM/OnboardingView3.swift
@@ -1,5 +1,5 @@
 //
-//  OnboardingView3.swift
+//  OnboardingView4.swift
 //  PIM
 //
 //  Created by 신정연 on 2023/09/16.
@@ -7,29 +7,18 @@
 
 import SwiftUI
 
-
-enum Pills: String, CaseIterable, Identifiable {
-    case 쎄스콘정
-    case 에이리스정
-    case 미니보라30
-    case 트리퀼라
-    var id: String { self.rawValue }
-}
-
 struct OnboardingView3: View {
     
-    let pills = ["쎄스콘정", "에이리스정", "미니보라30", "트리퀼라"]
-    @State var progressValue : Double = 0
-    @State private var selectedPill : String = ""
-    @State private var isPickerVisible = false
     @Environment(\.presentationMode) var presentationMode
+    
+    @State private var date = Date()
+    @State private var isNextButtonActive : Bool = false
     
     var body: some View {
         NavigationView{
             VStack {
-                HStack{
+                HStack {
                     Button(action: {
-                        // 이전 화면으로 이동
                         presentationMode.wrappedValue.dismiss()
                     }) {
                         Image(systemName: "chevron.left")
@@ -46,89 +35,33 @@ struct OnboardingView3: View {
                         .padding(.trailing, 25)
                     Spacer()
                 }
-                ProgressView(value: progressValue, total: 100)
+                ProgressView(value: 60, total: 100)
                     .progressViewStyle(LinearProgressViewStyle(tint: .red))
                     .padding(.bottom, 40)
-                    .onAppear{
-                        withAnimation(.linear(duration: 2)) { // 애니메이션 설정
-                            progressValue = 20
-                        }
-                    }
-                Text("어떤 약인가요?")
+                Text("현재 복용중인 패키지를\n시작한 날짜를 알려주세요")
+                    .frame(alignment: .center)
                     .font(.system(size: 24))
                     .fontWeight(.bold)
                     .padding(.bottom, 5)
-                Text("약 종류를 기록하면 내게 맞는 약을 찾기가 쉬워져요.")
+                Text("복용기간을 바탕으로 휴약기에는 알림이 울리지 않아요.")
                     .font(.body)
                     .foregroundColor(.gray)
                     .padding(.bottom, 40)
                 Image("character_onboarding_big")
                     .resizable()
                     .frame(width: 200, height: 200)
-                    .padding(.bottom, 50)
                 
-                //                Picker("약을 선택해주세요.", selection: $selected_pill){
-                //                    ForEach(0..<pills.count){
-                //                        Text(self.pills[$0])
-                //                    }
-                //                }
-                //                List{
-                //                    Picker("약을 선택해주세요.", selection: $selected_pill) {
-                //                        Text("쎄스콘정").tag(Pills.쎄스콘정)
-                //                        Text("미니보라30").tag(Pills.미니보라30)
-                //                        Text("에이리스정").tag(Pills.에이리스정)
-                //                        Text("트리퀼라").tag(Pills.트리퀼라)
-                //                    }
-                //                    .pickerStyle(.inline)
-                //                }
-                Button(action: {
-                    isPickerVisible.toggle()
-                }) {
-                    HStack {
-                        Text("약을 선택해주세요")
-                            .foregroundColor(.black)
-                            .font(.headline)
-                            .padding()
-//                            .background(Color.white)
-                            .cornerRadius(20)
-                        Spacer()
-                        Image(systemName: "chevron.down")
-                            .foregroundColor(Color.black)
-                            .padding(.trailing, 10)
-                    }
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 12)
-                            .stroke(Color.gray, lineWidth: 2)
-                    )
-                    
-                        
-                }
-                .padding(.leading, 20)
-                .padding(.trailing, 20)
-                .padding(.bottom, 10)
-                
-                if isPickerVisible {
-                    Picker("약을 선택해주세요", selection: $selectedPill) {
-                        ForEach(pills, id: \.self) { pill in
-                            Text(pill)
-                                .tag(pill)
-                                .font(.system(size: 18))
-                                .foregroundColor(selectedPill == pill ? .black : .gray)
-                                .fontWeight(selectedPill == pill ? .bold : .medium)
-                                .padding()
-//                                .background(
-//                                    RoundedRectangle(cornerRadius: 12)
-//                                        .frame(width: UIScreen.main.bounds.width * 0.8, height: 30)
-//                                        .foregroundColor(selectedPill == pill ? .gray : .clear)
-//                                )
-                        }
-                    }
-                    
-                    .pickerStyle(.inline)
-                    .transition(.slide) // Picker를 부드럽게 보이도록 애니메이션 추가
-                }
+                DatePicker(
+                    "",
+                    selection: $date,
+                    displayedComponents: [.date]
+                )
+                .datePickerStyle(.wheel)
+                .environment(\.locale, .init(identifier: "ko_KR"))
+               
                 Spacer()
-                NavigationLink(destination: OnboardingView4()) {
+                
+                NavigationLink(destination: OnboardingView4(selectedTime: .constant(Date()))) {
                     Text("선택했어요")
                         .font(.system(size: 20))
                         .fontWeight(.bold)
@@ -142,7 +75,7 @@ struct OnboardingView3: View {
         }
         .navigationBarBackButtonHidden(true)
     }
-    
+        
 }
 
 struct OnboardingView3_Previews: PreviewProvider {
