+++ PIM/PIM/OnboardingView4.swift
@@ -1,5 +1,5 @@
 //
-//  OnboardingView4.swift
+//  OnboardingView5.swift
 //  PIM
 //
 //  Created by 신정연 on 2023/09/16.
@@ -10,14 +10,16 @@ import SwiftUI
 struct OnboardingView4: View {
     
     @Environment(\.presentationMode) var presentationMode
-    @State private var date = Date()
+
+    @Binding var selectedTime: Date
+    @State private var isMainViewActive = false
+    let notificationManager = LocalNotificationManager()
     
     var body: some View {
         NavigationView{
             VStack {
-                HStack{
+                HStack {
                     Button(action: {
-                        // 이전 화면으로 이동
                         presentationMode.wrappedValue.dismiss()
                     }) {
                         Image(systemName: "chevron.left")
@@ -34,32 +36,75 @@ struct OnboardingView4: View {
                         .padding(.trailing, 25)
                     Spacer()
                 }
-                ProgressView(value: 60, total: 100)
+                ProgressView(value: 80, total: 100)
                     .progressViewStyle(LinearProgressViewStyle(tint: .red))
                     .padding(.bottom, 40)
-                Text("현재 복용중인 패키지를\n시작한 날짜를 알려주세요")
+                Text("몇 시에 약을 먹나요?")
                     .frame(alignment: .center)
                     .font(.system(size: 24))
                     .fontWeight(.bold)
                     .padding(.bottom, 5)
-                Text("복용기간을 바탕으로 휴약기에는 알림이 울리지 않아요.")
+                Text("선택한 복용 시간을 바탕으로 알림이 울려요.")
                     .font(.body)
                     .foregroundColor(.gray)
                     .padding(.bottom, 40)
                 Image("character_onboarding_big")
                     .resizable()
                     .frame(width: 200, height: 200)
+                    .padding(.bottom, 50)
                 
                 DatePicker(
                     "",
-                    selection: $date,
-                    displayedComponents: [.date]
+                    selection: $selectedTime,
+                    displayedComponents: [.hourAndMinute]
                 )
                 .datePickerStyle(.wheel)
                 .environment(\.locale, .init(identifier: "ko_KR"))
-               
+                .onChange(of: selectedTime) { newValue in
+                    print("Selected time changed to \(newValue)")
+                }
+                
                 Spacer()
-                NavigationLink(destination: OnboardingView5(selectedTime: $date)) {
+//                NavigationLink(destination: MainView()) {
+//                    Text("선택했어요")
+//                        .font(.system(size: 20))
+//                        .fontWeight(.bold)
+//                        .foregroundColor(Color.black)
+//                }
+//                .frame(width: UIScreen.main.bounds.width)
+//                .padding(.top, 40)
+//                .padding(.bottom, 10)
+//                .background(Color.gray)
+//                .contentShape(Rectangle())
+//                .onTapGesture {
+//                    print("Selected time: \(selectedTime)")
+//                    let calendar = Calendar.current
+//                    let hour = calendar.component(.hour, from: selectedTime)
+//                    let minute = calendar.component(.minute, from: selectedTime)
+//                    let selectedTimeString = "\(hour):\(minute)"
+//                    notificationManager.addNotification(title: "약 먹을 시간: \(selectedTimeString)")
+//                    UserDefaults.standard.set(selectedTime, forKey: "SelectedTime") // 사용자가 선택한 시간을 UserDefaults에 저장
+//                    notificationManager.schedule()
+//                    print("alert, at onboarding5: \(selectedTime)\n")
+//                }
+                
+                NavigationLink(
+                    destination: MainView().navigationBarHidden(true),
+                    isActive: $isMainViewActive) {
+                    EmptyView()
+                }
+                Button(action: {
+                    print("Button tapped")
+                    let calendar = Calendar.current
+                    let hour = calendar.component(.hour, from: selectedTime)
+                    let minute = calendar.component(.minute, from: selectedTime)
+                    let selectedTimeString = "\(hour):\(minute)"
+                    notificationManager.addNotification(title: "PIM")
+                    UserDefaults.standard.set(selectedTime, forKey: "SelectedTime")
+                    notificationManager.schedule()
+                    print("alert, at onboarding5: \(selectedTime)\n")
+                    isMainViewActive = true
+                }) {
                     Text("선택했어요")
                         .font(.system(size: 20))
                         .fontWeight(.bold)
@@ -78,6 +123,6 @@ struct OnboardingView4: View {
 
 struct OnboardingView4_Previews: PreviewProvider {
     static var previews: some View {
-        OnboardingView4()
+        OnboardingView4(selectedTime: .constant(Date()))
     }
 }
