+++ PIM/PIM/PIMApp.swift
@@ -15,7 +15,7 @@ struct PIMApp: App {
         WindowGroup {
 //            ContentView()
 //                .environment(\.managedObjectContext, persistenceController.container.viewContext)
-            OnboardingView1()
+            OnboardingMainView()
         }
     }
 }
